requirements.txt
pandas
requests
sqlalchemy
alembic
psycopg2
python-dotenv